refactor: split release command family
